@@ -4,19 +4,19 @@
   "private": true,
   "license": "MIT",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.3.0",
-    "@testing-library/user-event": "^14.3.0",
-    "immer": "^9.0.15",
-    "opencv-react": "^0.0.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-hook-form": "^7.33.1",
-    "react-icons": "^4.4.0",
+    "@testing-library/jest-dom": "^6.1.4",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.5.1",
+    "immer": "^10.0.3",
+    "opencv-react": "^0.0.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.47.0",
+    "react-icons": "^4.11.0",
     "react-scripts": "5.0.1",
-    "react-tooltip": "^4.2.21",
-    "sanitize-html": "^2.7.1",
-    "zustand": "^4.0.0"
+    "react-tooltip": "^5.21.5",
+    "sanitize-html": "^2.11.0",
+    "zustand": "^4.4.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -43,8 +43,8 @@
     ]
   },
   "devDependencies": {
-    "autoprefixer": "^10.4.7",
+    "autoprefixer": "^10.4.16",
     "postcss": "^8.4.31",
-    "tailwindcss": "^3.1.6"
+    "tailwindcss": "^3.3.3"
   }
 }
